+++ CHANGELOG.rst
@@ -2,7 +2,7 @@ This document contains the FSL-MRS release history in reverse chronological orde
 
 2.2.0 (WIP)
 -----------------------------
-- `fsl_dynmrs` can now handle MRSI data.
+- `fsl_dynmrs` can now handle MRSI data. This requires the use of `fsl_sub`, which is now a dependency.
 
 2.1.7 (Friday 7th July 2023)
 ----------------------------